+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- A class that handles the binaries filenames and its tests. This abstraction will help in future refactoring. By [@fschuch](https://github.com/fschuch).
+- A class that handles the binaries filenames and its tests, by [@fschuch](https://github.com/fschuch).
+- A class to handle the coordinates and its tests, so it can be moved out of the parameters class, by [@fschuch](https://github.com/fschuch).
 
 ## [0.1.11] - 2021-02-12